refactor(model): Refactor Errors and Exceptions to a Model Package
