+++ src/main/kotlin/com/expediagroup/sdk/core/config/Config.kt
@@ -16,7 +16,7 @@
 package com.expediagroup.sdk.core.config
 
 import com.expediagroup.sdk.core.config.provider.ConfigurationData
-import com.expediagroup.sdk.core.exception.ConfigurationException
+import com.expediagroup.sdk.core.model.exception.ConfigurationException
 
 /**
  * @property configurationData deserialized configuration data which needs to be parsed
