+++ src/main/kotlin/com/expediagroup/sdk/core/config/ConfigurationDefinition.kt
@@ -15,7 +15,7 @@
  */
 package com.expediagroup.sdk.core.config
 
-import com.expediagroup.sdk.core.exception.ConfigurationException
+import com.expediagroup.sdk.core.model.exception.ConfigurationException
 
 class ConfigurationDefinition {
