+++ src/main/kotlin/com/expediagroup/sdk/core/config/provider/FileConfigurationProvider.kt
@@ -15,7 +15,7 @@
  */
 package com.expediagroup.sdk.core.config.provider
 
-import com.expediagroup.sdk.core.exception.ConfigurationException
+import com.expediagroup.sdk.core.model.exception.ConfigurationException
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.BufferedReader
