+++ src/main/kotlin/com/expediagroup/sdk/core/model/exception/BaseException.kt
@@ -13,6 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.core.exception
+package com.expediagroup.sdk.core.model.exception
 
-class ConfigurationException(message: String?) : BaseException(message)
+internal open class BaseException : RuntimeException {
+    constructor(message: String?) : super(message)
+    constructor(message: String?, cause: Throwable?) : super(message, cause)
+}
