+++ src/main/kotlin/com/expediagroup/sdk/core/model/exception/ServiceException.kt
@@ -13,10 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.core.exception
+package com.expediagroup.sdk.core.model.exception
 
-open class BaseException : RuntimeException {
-    constructor(message: String?) : super(message)
-    constructor(message: String?, cause: Throwable?) : super(message, cause)
-    constructor(cause: Throwable?) : super(cause)
-}
+import com.expediagroup.sdk.core.model.error.Error
+import io.ktor.http.HttpStatusCode
+
+internal class ServiceException(
+    errorCode: HttpStatusCode,
+    error: Error
+) : BaseException("[${errorCode.value}] $error")
