+++ src/test/kotlin/com/expediagroup/sdk/core/config/provider/FileConfigurationProviderTest.kt
@@ -16,7 +16,7 @@
 package com.expediagroup.sdk.core.config.provider
 
 import com.expediagroup.sdk.core.constant.Constant.EMPTY_STRING
-import com.expediagroup.sdk.core.exception.ConfigurationException
+import com.expediagroup.sdk.core.model.exception.ConfigurationException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
